arun_backend/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
chromadb
sentence-transformers
alembic
numpy
scikit-learn
pandas
python-jose[cryptography]
passlib[bcrypt]
python-multipart
qdrant-client
hypothesis
